Capitalise the first letter of each word
Updated the return value to Capitalise the first letter of each word. When you create a macro which "adds tags", even if the tag you want to add (using the macro) contains any capital letters then when you actually use the macro, those capitals are changed to lowercase. This update ensures that your custom field data (such as a solution) is clean. For example "provision call plan" will show as "Provision Call Plan".